hwmon: Add basic MSI ATX PSU monitoring support
This patch adds basic support for MSI ATX PSUs with
a custom USB HID interface. The driver is loosely
based on the existing corsair-psu driver,
but with a few changes to support the different
HID protocol.

The driver is tested on the MSI MEG Ai1300P PCIE5 PSU,
with the exception of alarms as they would also
require the PSU to stay live after a fault condition.
It is therefore unclear how the PSU actually
implements this feature.

The driver is not yet complete, and is missing the
following features:
- Fan speed setting (auto or fixed PWM)
- OCP mode setting (multi or single rail)

Signed-off-by: Nimish Telang <nimish@telang.net>